+++ app/controllers/forest_liana/resources_controller.rb
@@ -62,6 +62,13 @@ module ForestLiana
       rescue ForestLiana::Errors::LiveQueryError => error
         render json: { errors: [{ status: 422, detail: error.message }] },
           status: :unprocessable_entity, serializer: nil
+      rescue ForestLiana::Errors::ExpectedError => error
+        error.display_error
+        error_data = JSONAPI::Serializer.serialize_errors([{
+          status: error.error_code,
+          detail: error.message
+        }])
+        render(serializer: nil, json: error_data, status: error.status)
       rescue => error
         FOREST_LOGGER.error "Records Index Count error: #{error}\n#{format_stacktrace(error)}"
         internal_server_error