+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- Collection - Avoid crash when filtering on a bad association or a bad field in resource getter.
+- Technical - Add test for resource getter to check error handling.
 
 ## RELEASE 2.14.0 - 2018-10-25
 ### Added